requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# API Clients
httpx==0.25.2
# aiohttp>=3.9.1  # Commented out due to compilation issues on Python 3.13 + macOS ARM64
openai==1.3.7
groq>=0.4.0  # Groq API client for openai/gpt-oss-120b model

# Image Processing
pillow==10.4.0
pytesseract==0.3.10

# Data Processing
# numpy>=1.26.4  # Commented out due to compilation issues on Python 3.13 + macOS ARM64
# pandas==2.1.3   # Depends on numpy, commented out
pydantic>=2.5.0  # Updated to allow newer versions compatible with Python 3.13
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
redis>=5.0.1
# asyncio==3.4.3  # Built-in module, no need to specify
# uuid==1.30       # Built-in module, no need to specify
psutil>=5.9.6

# Logging and Monitoring
structlog>=23.2.0
sentry-sdk>=1.38.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
# httpx-mock>=0.27.0  # Package not found, consider using pytest-httpx or respx instead

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Math and Science
sympy>=1.12
# matplotlib==3.8.2  # May have compilation issues on Python 3.13 + macOS ARM64
# scipy==1.11.4     # Depends on numpy, will fail

# UI
gradio>=4.0.0
